+++ spec/rubocop/cop/rails/file_path_spec.rb
@@ -40,4 +40,22 @@ describe RuboCop::Cop::Rails::FilePath do
       expect(cop.offenses.size).to eq(1)
     end
   end
+
+  context 'Rails.root is used as a method argument' do
+    let(:source) { 'foo(bar(File.join(Rails.root, "app", "models")))' }
+
+    it 'registers an offense once' do
+      inspect_source(cop, source)
+      expect(cop.offenses.size).to eq(1)
+    end
+  end
+
+  context 'Rails.root.join used as an argument' do
+    let(:source) { 'foo(Rails.root.join(\'app/models\'))' }
+
+    it 'registers an offense once' do
+      inspect_source(cop, source)
+      expect(cop.offenses.size).to eq(1)
+    end
+  end
 end